:zap: int型主键改为long，全面采用snowflake算法；优化xss过滤判断条件

+++ blade-core-tool/src/main/java/org/springblade/core/tool/node/ForestNode.java
@@ -34,7 +34,7 @@ public class ForestNode extends BaseNode {
 	 */
 	private Object content;
 
-	public ForestNode(Integer id, Integer parentId, Object content) {
+	public ForestNode(Long id, Long parentId, Object content) {
 		this.id = id;
 		this.parentId = parentId;
 		this.content = content;
