+++ blade-core-tool/src/main/java/org/springblade/core/tool/node/ForestNodeManager.java
@@ -34,7 +34,7 @@ public class ForestNodeManager<T extends INode> {
 	/**
 	 * 森林的父节点ID
 	 */
-	private List<Integer> parentIds = new ArrayList<>();
+	private List<Long> parentIds = new ArrayList<>();
 
 	public ForestNodeManager(List<T> items) {
 		list = items;
@@ -46,9 +46,9 @@ public class ForestNodeManager<T extends INode> {
 	 * @param id 节点ID
 	 * @return 对应的节点对象
 	 */
-	public INode getTreeNodeAT(int id) {
+	public INode getTreeNodeAT(Long id) {
 		for (INode forestNode : list) {
-			if (forestNode.getId() == id) {
+			if (forestNode.getId().longValue() == id.longValue()) {
 				return forestNode;
 			}
 		}
@@ -60,7 +60,7 @@ public class ForestNodeManager<T extends INode> {
 	 *
 	 * @param parentId
 	 */
-	public void addParentId(Integer parentId) {
+	public void addParentId(Long parentId) {
 		parentIds.add(parentId);
 	}
