+++ blade-core-tool/src/main/java/org/springblade/core/tool/node/TreeNode.java
@@ -30,8 +30,8 @@ public class TreeNode extends BaseNode {
 
 	private String title;
 
-	private Integer key;
+	private Long key;
 
-	private Integer value;
+	private Long value;
 
 }
